refactor(ai): split AiMessageList into ErrorGuideCard, AiCandidateCards, AiJudgeResult (#1279)
Extract ErrorGuideCard + AiPanelErrorDisplay to ErrorGuideCard.tsx,
AiCandidateCards to AiCandidateCards.tsx, and AiJudgeResult to
AiJudgeResult.tsx. Reduces AiMessageList from 488 to 234 lines.
